requirements.txt
tensorflow-gpu
pillow
moviepy